更新

+++ library/src/com/nostra13/universalimageloader/cache/disc/LimitedDiscCache.java
@@ -15,75 +15,88 @@
  *******************************************************************************/
 package com.nostra13.universalimageloader.cache.disc;
 
+import com.nostra13.universalimageloader.cache.disc.naming.FileNameGenerator;
+import com.nostra13.universalimageloader.core.DefaultConfigurationFactory;
+
 import java.io.File;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
-
-import com.nostra13.universalimageloader.cache.disc.naming.FileNameGenerator;
-import com.nostra13.universalimageloader.core.DefaultConfigurationFactory;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Abstract disc cache limited by some parameter. If cache exceeds specified limit then file with the most oldest last
  * usage date will be deleted.
- * 
+ *
  * @author Sergey Tarasevich (nostra13[at]gmail[dot]com)
- * @since 1.0.0
  * @see BaseDiscCache
  * @see FileNameGenerator
+ * @since 1.0.0
  */
 public abstract class LimitedDiscCache extends BaseDiscCache {
 
-	private int cacheSize = 0;
+	private static final int INVALID_SIZE = -1;
 
-	private int sizeLimit;
+	private final AtomicInteger cacheSize;
+
+	private final int sizeLimit;
 
 	private final Map<File, Long> lastUsageDates = Collections.synchronizedMap(new HashMap<File, Long>());
 
 	/**
-	 * @param cacheDir Directory for file caching. <b>Important:</b> Specify separate folder for cached files. It's
-	 *            needed for right cache limit work.
+	 * @param cacheDir  Directory for file caching. <b>Important:</b> Specify separate folder for cached files. It's
+	 *                  needed for right cache limit work.
 	 * @param sizeLimit Cache limit value. If cache exceeds this limit then file with the most oldest last usage date
-	 *            will be deleted.
+	 *                  will be deleted.
 	 */
 	public LimitedDiscCache(File cacheDir, int sizeLimit) {
 		this(cacheDir, DefaultConfigurationFactory.createFileNameGenerator(), sizeLimit);
 	}
 
 	/**
-	 * @param cacheDir Directory for file caching. <b>Important:</b> Specify separate folder for cached files. It's
-	 *            needed for right cache limit work.
+	 * @param cacheDir          Directory for file caching. <b>Important:</b> Specify separate folder for cached files. It's
+	 *                          needed for right cache limit work.
 	 * @param fileNameGenerator Name generator for cached files
-	 * @param sizeLimit Cache limit value. If cache exceeds this limit then file with the most oldest last usage date
-	 *            will be deleted.
+	 * @param sizeLimit         Cache limit value. If cache exceeds this limit then file with the most oldest last usage date
+	 *                          will be deleted.
 	 */
 	public LimitedDiscCache(File cacheDir, FileNameGenerator fileNameGenerator, int sizeLimit) {
 		super(cacheDir, fileNameGenerator);
 		this.sizeLimit = sizeLimit;
+		cacheSize = new AtomicInteger();
 		calculateCacheSizeAndFillUsageMap();
 	}
 
 	private void calculateCacheSizeAndFillUsageMap() {
-		int size = 0;
-		File[] cachedFiles = cacheDir.listFiles();
-		for (File cachedFile : cachedFiles) {
-			size += getSize(cachedFile);
-			lastUsageDates.put(cachedFile, cachedFile.lastModified());
-		}
-		cacheSize = size;
+		new Thread(new Runnable() {
+			@Override
+			public void run() {
+				int size = 0;
+				File[] cachedFiles = cacheDir.listFiles();
+				if (cachedFiles != null) { // rarely but it can happen, don't know why
+					for (File cachedFile : cachedFiles) {
+						size += getSize(cachedFile);
+						lastUsageDates.put(cachedFile, cachedFile.lastModified());
+					}
+					cacheSize.set(size);
+				}
+			}
+		}).start();
 	}
 
 	@Override
 	public void put(String key, File file) {
 		int valueSize = getSize(file);
-		while (cacheSize + valueSize > sizeLimit) {
+		int curCacheSize = cacheSize.get();
+
+		while (curCacheSize + valueSize > sizeLimit) {
 			int freedSize = removeNext();
-			if (freedSize == 0) break; // cache is empty (have nothing to delete)
-			cacheSize -= freedSize;
+			if (freedSize == INVALID_SIZE) break; // cache is empty (have nothing to delete)
+			curCacheSize = cacheSize.addAndGet(-freedSize);
 		}
-		cacheSize += valueSize;
+		cacheSize.addAndGet(valueSize);
 
 		Long currentTime = System.currentTimeMillis();
 		file.setLastModified(currentTime);
@@ -104,16 +117,15 @@ public abstract class LimitedDiscCache extends BaseDiscCache {
 	@Override
 	public void clear() {
 		lastUsageDates.clear();
-		cacheSize = 0;
+		cacheSize.set(0);
 		super.clear();
 	}
 
 	/** Remove next file and returns it's size */
 	private int removeNext() {
 		if (lastUsageDates.isEmpty()) {
-			return 0;
+			return INVALID_SIZE;
 		}
-
 		Long oldestUsage = null;
 		File mostLongUsedFile = null;
 		Set<Entry<File, Long>> entries = lastUsageDates.entrySet();
@@ -132,9 +144,16 @@ public abstract class LimitedDiscCache extends BaseDiscCache {
 			}
 		}
 
-		int fileSize = getSize(mostLongUsedFile);
-		if (mostLongUsedFile.delete()) {
-			lastUsageDates.remove(mostLongUsedFile);
+		int fileSize = 0;
+		if (mostLongUsedFile != null) {
+			if (mostLongUsedFile.exists()) {
+				fileSize = getSize(mostLongUsedFile);
+				if (mostLongUsedFile.delete()) {
+					lastUsageDates.remove(mostLongUsedFile);
+				}
+			} else {
+				lastUsageDates.remove(mostLongUsedFile);
+			}
 		}
 		return fileSize;
 	}
